# Conflicts:
#	src/main/java/br/com/utfpr/gerenciamento/server/dto/dashboards/DashboardEmprestimoCountRangeResponseDto.java
#	src/main/java/br/com/utfpr/gerenciamento/server/security/WebSecurity.java
#	src/main/java/br/com/utfpr/gerenciamento/server/service/impl/DashboardServiceImpl.java
#	src/main/java/br/com/utfpr/gerenciamento/server/service/impl/EmprestimoServiceImpl.java
#	src/main/java/br/com/utfpr/gerenciamento/server/service/impl/UsuarioServiceImpl.java
#	src/main/resources/db/common/V2.3__Create_System_Config_Table.sql
#	src/main/resources/db/test/V2.3__Create_System_Config_Table.sql

+++ src/main/java/br/com/utfpr/gerenciamento/server/model/dashboards/DashboardEmprestimoCountRange.java
@@ -4,13 +4,19 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+/**
+ * DTO para contadores de empréstimos agregados por status.
+ *
+ * <p>Usa Long para evitar overflow em sistemas de longo prazo com muitos registros.
+ * Math.toIntExact() anterior poderia lançar ArithmeticException se count > 2.1B.
+ */
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class DashboardEmprestimoCountRange {
 
-  private Integer total;
-  private Integer emAndamento;
-  private Integer emAtraso;
-  private Integer finalizado;
+  private Long total;
+  private Long emAndamento;
+  private Long emAtraso;
+  private Long finalizado;
 }
